requirements.txt
Django==4.0.5
psycopg2==2.9.3
djangorestframework==3.13.1
django-phonenumber-field[phonenumbers]==6.3.0
django-rest-knox==4.2.0
python-dotenv==0.20.0